agents/price_comparison/requirements.txt
# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Database
redis>=5.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# AI/ML
langchain-core>=0.3.0
langchain>=0.3.0
langchain-community>=0.3.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
httpx>=0.25.0  # For async HTTP requests

# Web Crawler Dependencies
playwright>=1.42.0
crawl4ai>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0

# New dependencies
langchain==0.1.12
langchain-community==0.0.29
langchain-core==0.1.30
langchain-openai==0.0.8
loguru==0.7.2
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
redis==5.0.3 